CTR: don't change the stream length.
For the last iteration of the encryption loop, the "block" to encrypt
may not be a "full" block of blocksize length. As a result, we need to
truncate the keystream to match the partial block length before XOR'ing
them.

Before this patch, the stream bytes struct itself would be truncated for
the last iteration. It just doesn't "feel right" to hack the stream
directly, so this patch leave the stream as-is and use a truncated copy
for the XOR operation.