Fix test_bignum_to_bytes()
Before this patch, a completely random bytes buffer was used to be
converted to bignum and then back a to a bytes buffer again for
comparison. When the random buffer most significant byte was zero, this
byte was "lost" when converted to a number. This patch ensure that the
first byte is always non-zero.