hmac tweak.
Before this patch, when the key would be hashed, it would be then
truncated to L bytes. This patch only ensure that its length is L.

The output result of the provided H() function must always be of length
L. Thus, there is no need to try to truncate it and we can add a sanity
check to verify its length instead.